pyproject.toml
[project]
name = "projet2"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests (>=2.33.0,<3.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "tqdm (>=4.67.3,<5.0.0)"
]

[tool.poetry]
packages = [{include = "projet2", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
